@@ -9,26 +9,26 @@
     "isomorphic-fetch": "^2.2.1",
     "react-addons-test-utils": "^15.4.2",
     "react-container-query": "^0.8.1",
+    "react-scripts": "0.9.5",
     "redux-devtools": "^3.3.2",
-    "redux-devtools-log-monitor": "^1.2.0",
     "redux-devtools-dock-monitor": "^1.1.1",
-    "react-scripts": "0.8.5"
+    "redux-devtools-log-monitor": "^1.2.0"
   },
   "dependencies": {
     "classnames": "^2.2.5",
     "font-awesome": "^4.7.0",
     "immutable": "^3.8.1",
-    "nock": "^9.0.6",
-    "normalizr": "^2.3.0",
+    "nock": "^9.0.13",
+    "normalizr": "^3.2.2",
     "purecss": "^0.6.2",
-    "react": "^15.4.2",
-    "react-dom": "^15.4.2",
-    "react-redux": "^4.4.6",
-    "react-router": "^3.0.2",
+    "react": "15.5.4",
+    "react-dom": "15.5.4",
+    "react-redux": "^5.0.4",
+    "react-router": "^4.1.1",
     "react-router-redux": "^4.0.8",
     "redux": "^3.6.0",
-    "redux-logger": "^2.8.1",
-    "redux-mock-store": "^1.2.2",
+    "redux-logger": "^3.0.1",
+    "redux-mock-store": "^1.2.3",
     "redux-thunk": "^2.2.0",
     "reselect": "^2.5.4"
   },
