@@ -1,5 +1,6 @@
 //containers/Root.dev.js
-import React, { PropTypes } from 'react'
+import React from 'react'
+import PropTypes from 'prop-types'
 import { Provider } from 'react-redux'
 import routes from '../routes'
 import DevTools from './DevTools'
